multi-region-gke/app/database/requirements.txt
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-spanner==1.7.1
googleapis-common-protos==1.6.0
grpc-google-iam-v1==0.11.4
grpcio==1.23.0
grpcio-gcp==0.2.2
idna==2.8
lorem==0.1.1
protobuf==3.9.1
pyasn1==0.4.7
pyasn1-modules==0.2.6
pytz==2019.2
requests==2.22.0
rsa==4.0
six==1.12.0
urllib3==1.25.3
